Refactored

+++ graph_utils/canonical_graph_generator_test.cc
@@ -1,3 +1,4 @@
+// Unit tests for the canonical graph generator.
 
 #include "canonical_graph_generator.h"
 
@@ -14,12 +15,10 @@
 #include "gtest/gtest.h"
 #include "nauty_utils/nauty_wrapper.h"
 
-
 using std::set;
 using std::vector;
 using graph_utils::Graph;
 using nauty_utils::IsomorphismChecker;
-
 
 namespace graph_utils {
 namespace {
@@ -27,14 +26,14 @@ namespace {
 template <typename T>
 void ExpectVectorsEq(const vector<T> &v1, const vector<T> &v2) {
   ASSERT_EQ(v1.size(), v2.size());
-  for (uint i = 0; i < v1.size(); ++i) {
+  for (size_t i = 0; i < v1.size(); ++i) {
     EXPECT_EQ(v1[i], v2[i]);
   }
 }
 
 int CountConnectedGraphs(const vector<Graph *> &graphs) {
   int count = 0;
-  for (int i = 0; i < graphs.size(); ++i) {
+  for (size_t i = 0; i < graphs.size(); ++i) {
     if (graphs[i]->IsConnected()) {
       ++count;
     }
@@ -43,14 +42,13 @@ int CountConnectedGraphs(const vector<Graph *> &graphs) {
 }
 
 class AllGrapsAcceptable : public CanonicalGraphFilter {
- public:
-  virtual bool IsSubsetSafe(const Graph &g,
-                            const vector<int> &subset) const {
+public:
+  virtual bool IsSubsetSafe(const Graph &g, const vector<int> &subset) const {
     return true;
   }
 };
 
-}  // namespace
+} // namespace
 
 class CanonicalGraphGeneratorTest : public testing::Test {
 protected:
@@ -59,37 +57,38 @@ protected:
 };
 
 TEST_F(CanonicalGraphGeneratorTest, UpperObjects_EmptyGraphTest) {
-  vector<string> v({ "000", "000", "000" });
+  vector<string> v({"000", "000", "000"});
   Graph g(v);
   CanonicalGraphGenerator generator(g.size(), filter_.get());
   vector<Graph *> upper_obj;
   generator.GenerateUpperObjects(g, &upper_obj);
   ASSERT_EQ(7, upper_obj.size());
-  ASSERT_EQ(4, upper_obj[0]->size());  // The graphs are of order 4.
+  ASSERT_EQ(4, upper_obj[0]->size()); // The graphs are of order 4.
 }
 
 TEST_F(CanonicalGraphGeneratorTest, UpperObjects_TriangleGraphTest) {
-  vector<string> v({ "011", "101", "110" });
+  vector<string> v({"011", "101", "110"});
   Graph g(v);
   CanonicalGraphGenerator generator(g.size(), filter_.get());
   vector<Graph *> upper_obj;
   generator.GenerateUpperObjects(g, &upper_obj);
   ASSERT_EQ(3, upper_obj.size());
-  ASSERT_EQ(4, upper_obj[0]->size());  // The graphs are of order 4.
+  ASSERT_EQ(4, upper_obj[0]->size()); // The graphs are of order 4.
+  // Check each of the upper objects individually.
   {
-    vector<string> expected({ "0111", "1010", "1100", "1000" });
+    vector<string> expected({"0111", "1010", "1100", "1000"});
     vector<string> mat;
     upper_obj[0]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({ "0110", "1011", "1100", "0100" });
+    vector<string> expected({"0110", "1011", "1100", "0100"});
     vector<string> mat;
     upper_obj[1]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({ "0110", "1010", "1101", "0010" });
+    vector<string> expected({"0110", "1010", "1101", "0010"});
     vector<string> mat;
     upper_obj[2]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
@@ -97,47 +96,47 @@ TEST_F(CanonicalGraphGeneratorTest, UpperObjects_TriangleGraphTest) {
 }
 
 TEST_F(CanonicalGraphGeneratorTest, UpperObjects_LineGraphTest) {
-  vector<string> v({ "010", "101", "010" });
+  vector<string> v({"010", "101", "010"});
   Graph g(v);
   CanonicalGraphGenerator generator(g.size(), filter_.get());
   vector<Graph *> upper_obj;
   generator.GenerateUpperObjects(g, &upper_obj);
   ASSERT_EQ(6, upper_obj.size());
-  ASSERT_EQ(4, upper_obj[0]->size());  // The graphs are of order 4.
-
+  ASSERT_EQ(4, upper_obj[0]->size()); // The graphs are of order 4.
+  // Check each of the upper objects individually.
   {
-    vector<string> expected({ "0101", "1010", "0100", "1000" });
+    vector<string> expected({"0101", "1010", "0100", "1000"});
     vector<string> mat;
     upper_obj[0]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({ "0100", "1011", "0100", "0100" });
+    vector<string> expected({"0100", "1011", "0100", "0100"});
     vector<string> mat;
     upper_obj[1]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({ "0101", "1011", "0100", "1100" });
+    vector<string> expected({"0101", "1011", "0100", "1100"});
     vector<string> mat;
     upper_obj[2]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
 
   {
-    vector<string> expected({ "0100", "1010", "0101", "0010" });
+    vector<string> expected({"0100", "1010", "0101", "0010"});
     vector<string> mat;
     upper_obj[3]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({ "0101", "1010", "0101", "1010" });
+    vector<string> expected({"0101", "1010", "0101", "1010"});
     vector<string> mat;
     upper_obj[4]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({ "0100", "1011", "0101", "0110" });
+    vector<string> expected({"0100", "1011", "0101", "0110"});
     vector<string> mat;
     upper_obj[5]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
@@ -145,32 +144,33 @@ TEST_F(CanonicalGraphGeneratorTest, UpperObjects_LineGraphTest) {
 }
 
 TEST_F(CanonicalGraphGeneratorTest, LowerObjectsTest) {
-  vector<string> v({ "0100", "1011", "0101", "0110" });
+  vector<string> v({"0100", "1011", "0101", "0110"});
   Graph g(v);
   CanonicalGraphGenerator generator(g.size(), filter_.get());
   vector<Graph *> lower_obj;
   generator.GenerateLowerObjects(g, &lower_obj);
   ASSERT_EQ(4, lower_obj.size());
+  // Check each of the lower objects individually.
   {
-    vector<string> expected({ "011", "101", "110" });
+    vector<string> expected({"011", "101", "110"});
     vector<string> mat;
     lower_obj[0]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({ "000", "001", "010" });
+    vector<string> expected({"000", "001", "010"});
     vector<string> mat;
     lower_obj[1]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({"010", "101", "010" });
+    vector<string> expected({"010", "101", "010"});
     vector<string> mat;
     lower_obj[2]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
   }
   {
-    vector<string> expected({ "010", "101", "010" });
+    vector<string> expected({"010", "101", "010"});
     vector<string> mat;
     lower_obj[3]->GetAdjMatrix(&mat);
     ExpectVectorsEq<string>(expected, mat);
